WIP: A bit more functional
