+++ .gitignore
@@ -1,4 +1,8 @@
 bower_components
-node_modules
+node_modules/*
+lib/*
 *.iml
 .idea
+bundle.js
+css/materialize.min.css
+css/starter-template.css
