+++ js/app.js
@@ -1,0 +1,5 @@
+// var $ = require("jquery");
+var canvas = require("./canvas.js");
+
+// export canvas functions to global namespace
+$.extend(window, canvas);
